Updated credit cards ui
